fix dtype for target vectors

+++ src/main.py
@@ -340,7 +340,7 @@ def main(args):
             train_fake.extend(train_real)
             train_array=torch.Tensor(train_fake)
             train_target_fake.extend(train_target_real)
-            train_targets=torch.Tensor(train_target_fake)
+            train_targets=torch.Tensor(train_target_fake, dtype=torch.long)
             train_set = TensorDataset(train_array,train_targets)
 
             #BUILD VAL SET
@@ -353,7 +353,7 @@ def main(args):
             val_fake.extend(val_real)
             val_array=torch.Tensor(val_fake)
             val_target_fake.extend(val_target_real)
-            val_targets=torch.Tensor(val_target_fake)
+            val_targets=torch.Tensor(val_target_fake, dtype=torch.long)
             val_set = TensorDataset(val_array, val_targets)
         else:
             train_set= ImageFolderWithPaths(args.data_root,transform=transform)
@@ -382,7 +382,7 @@ def main(args):
             test_fake.extend(test_real)
             test_array=torch.Tensor(test_fake)
             test_target_fake.extend(test_target_real)
-            test_targets=torch.Tensor(test_target_fake)
+            test_targets=torch.Tensor(test_target_fake, dtype=torch.long)
             te_dataset = TensorDataset(test_array,test_targets)
         else:
             te_dataset=ImageFolderWithPaths(args.data_root,transform=transform)
